% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_simulation.R
\name{compute_distance_covariance_matrix}
\alias{compute_distance_covariance_matrix}
\title{compute_distance_covariance_matrix}
\usage{
compute_distance_covariance_matrix(sf_object, l)
}
\arguments{
\item{sf_object}{Object of class \code{sf}}

\item{l}{Smoothness parameter: bigger \code{l} represents less correlation between
neighboring points.}
}
\value{
a matrix representing the covariance between points at different distances.
}
\description{
Given an sf object, compute a covariance matrix based on distances between
centroids using an exponential kernel.
}
