% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_simulation.R
\name{estimate_fine_grid}
\alias{estimate_fine_grid}
\title{estimate_fine_grid}
\usage{
estimate_fine_grid(fine_grid, coarse_grid, covariance_matrix, lambda)
}
\arguments{
\item{fine_grid}{A finer grid which we would like to estimate values for}

\item{coarse_grid}{A coarser grid whose values are known and in a \code{value} field}

\item{covariance_matrix}{A covariance matrix with one row/col per element of fine_grid}

\item{lambda}{A scaling parameter controlling the regularization}
}
\description{
Uses the values associated with geometries in a coarse grid to estimate values
for geometries in a finer grid. This is done using regularization that assumes
a gaussian prior with a given covariance structure.#'
}
