% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_simulation.R
\name{generate_boston_map}
\alias{generate_boston_map}
\title{generate_boston_map}
\usage{
generate_boston_map(plot = FALSE)
}
\arguments{
\item{plot}{A boolean value indicating whether or not to plot the resulting \code{sf} object of Boston.}
}
\value{
An \code{sf} object containing geometries for 5 counties in or near Boston.
}
\description{
generate_boston_map
}
\examples{

generate_boston_map(plot = TRUE)

}
