% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_simulation.R
\name{generate_coarse_grid_value}
\alias{generate_coarse_grid_value}
\title{generate_coarse_grid_value}
\usage{
generate_coarse_grid_value(fine_grid, coarse_grid)
}
\arguments{
\item{fine_grid}{A sf object thought to have smaller geometries; requires a \code{value} field}

\item{coarse_grid}{A sf object thought to have larger geometries}
}
\value{
An sf object equal to \code{coarse_grid} with a \code{value} field attached.
}
\description{
Given an sf object associated with a \code{value} and a second sf object, compute a value
for the second sf object based on the first objects' values over intersections.
}
