% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_simulation.R
\name{generate_gaussian_random_field}
\alias{generate_gaussian_random_field}
\title{generate_gaussian_random_field}
\usage{
generate_gaussian_random_field(sf_object, l = 1, plot = FALSE, ...)
}
\arguments{
\item{l}{Larger l means slower decay in covarinace (higher covariance between nearby points)}

\item{...}{}
}
\description{
generate_gaussian_random_field
}
\examples{

boston_map <- generate_boston_map()
gaussian_random_field <- generate_gaussian_random_field(boston_map,l = 1,n = c(30,30), theta = -pi/4, plot = TRUE)

}
