% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_simulation.R
\name{generate_grid}
\alias{generate_grid}
\title{generate_grid}
\usage{
generate_grid(sf_object, theta = 0, plot = FALSE, ...)
}
\arguments{
\item{sf_object}{An sf object over which to make the grid}

\item{plot}{A logical value representing whether or not to plot the result}

\item{...}{}
}
\description{
Given an sf object, generate an underlying grid sf object. Options exist to rotate the grid.
}
\examples{

boston_map <- generate_boston_map()
generate_grid(boston_map,plot = TRUE,n=c(10,10),theta = -pi/4)

}
